Separate into "route_intermid".
Addresses in issue #2.
