+++ single/main.cpp
@@ -1,5 +1,6 @@
 #include "./main.hpp"
 #include "./route.hpp"
+#include "./route_intermid.hpp"
 #include "./fixflag.hpp"
 
 /*******************************************************/
