+++ single/route_intermid.hpp
@@ -1,0 +1,11 @@
+/*******************************************************/
+/**  ルーティング (中間ポート利用) に関する関数       **/
+/*******************************************************/
+
+#ifndef __ROUTE_INTERMID_HPP__
+#define __ROUTE_INTERMID_HPP__
+
+bool routingSourceToI(int trgt_line_id);
+bool routingIToSink(int trgt_line_id);
+
+#endif
